Added mhash column to care and treatment dockets

+++ src/contracts/DwapiCentral.Contracts/Ct/ILab.cs
@@ -16,7 +16,7 @@ namespace DwapiCentral.Contracts.Ct
         string? SampleType { get; set; }
         string? Reason { get; set; }
 
-        int Mhash { get; set; }
+        ulong Mhash { get; set; }
 
 
     }
