+++ src/ct/DwapiCentral.Ct.Domain/Models/PatientLaboratoryExtract.cs
@@ -11,7 +11,7 @@ namespace DwapiCentral.Ct.Domain.Models
         public string RecordUUID { get; set; }
         public int PatientPk { get; set; }
         public int SiteCode { get; set; }
-        public int Mhash { get; set; }
+        public ulong Mhash { get; set; }
         public int? VisitId { get; set; }
         public DateTime? OrderedByDate { get; set; }
         public DateTime? ReportedByDate { get; set; }
